Improve test efficiency and move Tests next to Code
