+++ .github/workflows/pyright.yml
@@ -28,11 +28,3 @@ jobs:
       - name: Run pyright
         run: uv run pyright
 
-      # - name: Run docker-compose
-      #   uses: hoverkraft-tech/compose-action@v2.0.1
-      #   with:
-      #     compose-file: "./docker-compose.yaml"
-
-      # - name: Run tests
-      #   # For example, using `pytest`
-      #   run: uv run pytest
