--- tests/odwr/__init__.py
@@ -1,7 +1,0 @@
-# =================================================================
-#
-# Copyright (c) 2025 Lincoln Institute of Land Policy
-#
-# Licensed under the MIT License.
-#
-# =================================================================
