+++ userCode/awqms/sta_generation_test.py
