+++ userCode/odwr/lib_test.py
@@ -12,7 +12,7 @@ import datetime
 from hypothesis import given, strategies as st
 import pytest
 
-from ..lib import wipe_datastreams, wipe_locations, wipe_things
+from test.lib import wipe_datastreams, wipe_locations, wipe_things
 from userCode.odwr.lib import (
     assert_valid_oregon_date,
     download_oregon_tsv,
