+++ userCode/xlsx/main_test.py
@@ -15,9 +15,11 @@ import pytest
 from userCode.env import API_BACKEND_URL
 from userCode.xlsx.lib import parse_xlsx
 
+test_data = Path(__file__).parent / "testdata"
+
 
 def test_parse_xlsx():
-    file = Path(__file__).parent / "IoW_Reccomended_Obs_Data_Elements.xlsx"
+    file = test_data / "IoW_Reccomended_Obs_Data_Elements.xlsx"
     xlsx = parse_xlsx(file)
     assert xlsx.dataSheet
     assert xlsx.metadataSheet
@@ -37,20 +39,20 @@ def test_frost_connection():
 
 
 def test_parse_bad_xlsx():
-    file = Path(__file__).parent / "empty.xlsx"
+    file = test_data / "empty.xlsx"
     with pytest.raises(AssertionError):
         parse_xlsx(file)
 
 
 def test_parse_xlsx_and_generate_sta():
-    file = Path(__file__).parent / "IoW_Reccomended_Obs_Data_Elements.xlsx"
+    file = test_data / "IoW_Reccomended_Obs_Data_Elements.xlsx"
     xlsx = parse_xlsx(file)
     sta_representation = xlsx.to_sta()
     assert sta_representation
 
 
 def test_insert_data_into_frost():
-    xlsx = parse_xlsx(Path(__file__).parent / "IoW_Reccomended_Obs_Data_Elements.xlsx")
+    xlsx = parse_xlsx(test_data / "IoW_Reccomended_Obs_Data_Elements.xlsx")
     sta_representation = xlsx.to_sta()
     xlsx.send_to_frost(sta_representation)
